+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Version 0.1.20
 
+### Changed
+
+- improved UI for habits and wishlist on mobile
+
 ### Fixed
 
 - fix pomodoro break timer from triggering completions